requirements.txt
sshkeyboard==2.3.1
pyserial
